requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.0
sqlparse==0.4.4
typing-extensions==4.10.0
